API list views: optimize latest transcription lookup
This is expensive enough in the new viewer that I
moved the latest transcription lookup from a 
subquery to a separate stage to avoid it being run
twice (once when the pagination counts the number
of pages, once when actually retrieving the 
records).

In simple testing, this shaves ~100ms off of the
response time.